chapter8/requirements-dev.txt
langchain-core==0.1.18
langchain-openai==0.0.5
langchain==0.1.5
langchain-community==0.0.17
python-dotenv==1.0.1
pinecone-client==3.0.2
urllib3==1.26.18
unstructured==0.10.15
pdf2image==1.17.0
pdfminer.six==20231228
